fix: satisfy pkg-file-count limits for config init
Move you config command wiring into pkg/cli/configinit and fold built-in factory init helpers into named_factory_layout.go so pkg/cli and pkg/config stay within the 15-file package limit.
